refactor openapi-utilities ahead of oas2 support
