+++ projects/openapi-utilities/src/openapi/__tests__/sourcemap-reader.test.ts
@@ -1,5 +1,5 @@
 // TODO get rid of fixture import so we don't have circular references
-import { parseOpenAPIWithSourcemap } from '../../../../../../openapi-io/src/parser/openapi-sourcemap-parser';
+import { parseOpenAPIWithSourcemap } from '@useoptic/openapi-io/build/parser/openapi-sourcemap-parser';
 import path from 'path';
 import { sourcemapReader } from '../sourcemap-reader';
 const util = require('util');
@@ -10,7 +10,7 @@ const fixture = async () => {
     path.resolve(
       path.join(
         __dirname,
-        '../../../../../../openapi-io/inputs/openapi3-with-references/external-multiple.yaml'
+        '../../../../openapi-io/inputs/openapi3-with-references/external-multiple.yaml'
       )
     )
   );
@@ -72,7 +72,7 @@ describe('reading sourcemaps', () => {
       path.resolve(
         path.join(
           __dirname,
-          '../../../../../../openapi-io/inputs/openapi3/001-ok-add-property-field.yaml'
+          '../../../../openapi-io/inputs/openapi3/001-ok-add-property-field.yaml'
         )
       )
     );
