+++ projects/openapi-utilities/src/openapi/constants.ts
