+++ projects/openapi-utilities/src/openapi/sdk/isType.ts
