+++ projects/openapi-utilities/src/openapi/sdk/types/index.ts
