+++ projects/openapi-utilities/src/openapi/sdk/types/location.ts
