+++ projects/openapi-utilities/src/openapi/sourcemap-reader.ts
