+++ projects/openapi-utilities/src/openapi/validation-schemas.ts
