+++ projects/openapi-utilities/src/openapi/validator.ts
