+++ projects/openapi-utilities/src/utilities/__tests__/group-changes.test.ts
@@ -1,16 +1,16 @@
 import {
   factsToChangelog,
-  OpenAPITraverser,
+  OpenAPI3Traverser,
   groupChangesAndRules,
 } from '../../index';
 import { openAPI as petStoreBase } from '../../examples/petstore-base';
 import { openAPI as petStoreUpdated } from '../../examples/petstore-updated';
 
 test('groupChangesAndRules diffs and groups changes between two open api files', () => {
-  const baseTraverser = new OpenAPITraverser();
+  const baseTraverser = new OpenAPI3Traverser();
   baseTraverser.traverse(petStoreBase);
 
-  const targetTraverser = new OpenAPITraverser();
+  const targetTraverser = new OpenAPI3Traverser();
   targetTraverser.traverse(petStoreUpdated);
 
   const baseFacts = [...baseTraverser.facts()];
